@@ -3,20 +3,27 @@ import nox
 
 @nox.session(reuse_venv=True)
 def lint(session):
-    session.install("ruff", "black")
-    session.run("ruff", "check", ".")
-    session.run("black", "--check", ".")
+    session.run("uv", "sync", "--dev", external=True)
+    session.run("uv", "run", "sh", "scripts/lint.sh", external=True)
 
 
 @nox.session(reuse_venv=True)
 def typecheck(session):
-    session.install("mypy")
-    # Only include specific test files to avoid old/deprecated tests
-    sources = ["src/stbsim", "tests/test_core.py", "tests/__init__.py"]
-    session.run("mypy", *sources)
+    session.run("uv", "sync", "--dev", external=True)
+    session.run("uv", "run", "sh", "scripts/typecheck.sh", external=True)
 
 
 @nox.session(reuse_venv=True)
 def tests(session):
-    session.install("pytest", "pytest-cov")
-    session.run("pytest", "--cov=src/stbsim", "tests")
+    session.run("uv", "sync", "--dev", external=True)
+    session.run(
+        "uv",
+        "run",
+        "pytest",
+        "--cov=stbsim",
+        "--cov-report=term-missing",
+        "--cov-report=html:reports/coverage_html",
+        "--cov-fail-under=80",
+        "tests",
+        external=True,
+    )
